+++ src/entities/report/ui/report-form-content.tsx
@@ -5,7 +5,7 @@ import { useSendReport } from '@/entities/report/model/use-send-report'
 import { BasicTextarea } from '@/shared/ui/basic-textarea'
 import { Button } from '@/shared/ui/button'
 import { toast } from '@/shared/lib/use-toast'
-import { useIOSKeyboardPatch } from '@/shared/lib/use-keyboard-height'
+import { useIOSKeyboardPatch } from '@/shared/lib/use-ios-keyboard-patch'
 
 interface ReportFormContentProps {
   handleClose: () => void
@@ -36,7 +36,7 @@ export const ReportFormContent = ({ handleClose }: ReportFormContentProps) => {
         placeholder='내용을 작성해주세요.'
         value={comment}
         onChange={handleChange}
-        className='min-h-[141px] resize-none rounded-[0.375rem] border-slate-200 bg-slate-100 p-4 text-[1rem] leading-4 tracking-[-0.02rem] text-slate-600 scrollbar-none placeholder:text-slate-300 tab:min-h-[166px] tab:text-base pc:min-h-[9.96rem] pc:p-[0.8rem] pc:text-[1.2rem] pc:text-base pc:leading-[1.2rem]'
+        className='scrollbar-none min-h-[141px] resize-none rounded-[0.375rem] border-slate-200 bg-slate-100 p-4 text-[1rem] leading-4 tracking-[-0.02rem] text-slate-600 placeholder:text-slate-300 tab:min-h-[166px] tab:text-base pc:min-h-[9.96rem] pc:p-[0.8rem] pc:text-[1.2rem] pc:text-base pc:leading-[1.2rem]'
       />
       <Button
         disabled={!comment}
