+++ src/pages/results/ui/custom-text-editor.tsx
@@ -5,10 +5,9 @@ import { VisuallyHidden } from '@radix-ui/react-visually-hidden'
 
 import { Button } from '@/shared/ui/button'
 import { Popover, PopoverContent, PopoverTrigger } from '@/shared/ui/popover'
-import { DialogHeader, DialogTitle } from '@/shared/ui/test-dialog'
+import { DialogHeader, DialogTitle, Dialog } from '@/shared/ui/dialog'
 import { CustomTextEditorContent } from './custom-text-editor-content'
 import EditIcon from '@/shared/ui/icon/icon-pencil-black.svg'
-import { Dialog } from '@/shared/ui/test-dialog'
 
 interface CustomTextEditorProps {
   children: ReactNode
@@ -48,27 +47,11 @@ export const CustomTextEditor = ({ children }: CustomTextEditorProps) => {
       </Popover>
 
       {/* mobile,tab */}
-      {/* <Dialog open={isDialogOpen} onOpenChange={setDialogOpen}>
-        <DialogTrigger asChild className='pc:hidden'>
-          {children}
-        </DialogTrigger>
-        <DialogContent className='max-w-[19.17rem] rounded-2xl bg-white p-[1.125rem] pt-[0.9375rem] tab:max-w-[22.5625rem]'>
-          <DialogHeader>
-            <DialogTitle>
-              <VisuallyHidden>대치어 직접 수정하기</VisuallyHidden>
-            </DialogTitle>
-            <CustomTextEditorTitle />
-          </DialogHeader>
-          <CustomTextEditorContent handleClose={handleDialogClose} />
-        </DialogContent>
-      </Dialog> */}
-
-      {/* mobile,tab */}
       <Dialog
         onOpenChange={setDialogOpen}
         open={isDialogOpen}
         DialogTriggerItem={children}
-        className='max-w-[19.17rem] rounded-2xl bg-white p-[1.125rem] pt-[0.9375rem] scrollbar-none tab:max-w-[22.5625rem]'
+        className='scrollbar-none max-w-[19.17rem] rounded-2xl bg-white p-[1.125rem] pt-[0.9375rem] tab:max-w-[22.5625rem]'
       >
         <DialogHeader>
           <DialogTitle>
