+++ src/shared/lib/use-keyboard-height.ts
@@ -1,4 +1,4 @@
-import { RefObject, useCallback, useEffect, useState } from 'react'
+import { useEffect, useState } from 'react'
 
 const useKeyboardHeight = () => {
   const [keyboardHeight, setKeyboardHeight] = useState(0)
@@ -50,31 +50,3 @@ const useKeyboardHeight = () => {
 }
 
 export { useKeyboardHeight }
-
-export const useIOSKeyboardPatch = (
-  ref: RefObject<HTMLInputElement | HTMLTextAreaElement>,
-  shouldApply = false,
-) => {
-  const forceRefocus = useCallback(() => {
-    const input = ref.current
-    if (!shouldApply || !input) return
-
-    input.blur()
-    setTimeout(() => {
-      input.focus()
-    }, 100)
-  }, [ref])
-
-  useEffect(() => {
-    const handleTouchOutside = (e: TouchEvent) => {
-      if (ref.current && !ref.current.contains(e.target as Node)) {
-        forceRefocus()
-      }
-    }
-
-    document.addEventListener('touchend', handleTouchOutside)
-    return () => {
-      document.removeEventListener('touchend', handleTouchOutside)
-    }
-  }, [forceRefocus])
-}
